Minor bug fixes

+++ larndsim/consts/detector.py
@@ -410,7 +410,7 @@ def set_detector_properties(detprop_file, pixel_file, response_file=None, i_modu
         LONG_DIFF = float(detprop.get('long_diff', LONG_DIFF))
         TRAN_DIFF = float(detprop.get('tran_diff', TRAN_DIFF))
 
-        MAX_RADIUS = float(detprop.get('max_radius', MAX_RADIUS))
+        MAX_RADIUS = int(detprop.get('max_radius', MAX_RADIUS))
 
         # Prepare neighbouring pixel distance for backtracking
         # Currently backtracking range is used to convert from segment base to pixel base
